Refactor BaatoDio to include KeyHashInterceptors and streamline query parameter handling
